Add session handler
