+++ src/ChallengeManagerInterface.php
@@ -25,38 +25,5 @@ interface ChallengeManagerInterface
      *
      * @internal
      */
-    public function useFromClientDataJSON(string $base64url): ?ChallengeInterface;
-}
-
-
-class SessionChallengeManager
-{
-    private const SESSION_KEY = 'passkey_challenge';
-
-    public function __construct()
-    {
-        // do later?
-        if (session_status() !== PHP_SESSION_ACTIVE) {
-            throw new \BadMethodCallException('Call session_start()');
-        }
-    }
-
-    public function createChallenge(): ChallengeInterface
-    {
-        $c = ExpiringChallenge::withLifetime(120);
-        $_SESSION[self::SESSION_KEY] = $c;
-        return $c;
-    }
-
-    /*
-    public function getActiveChallenge(): ?ChallengeInterface
-    {
-        if (!array_key_exists(self::SESSION_KEY, $_SESSION)) {
-            return null;
-        }
-        $challenge = $_SESSION[self::SESSION_KEY];
-        unset($_SESSION[self::SESSION_KEY]);
-        return $challenge;
-    }
-     */
+    public function useFromClientDataJSON(string $base64Url): ?ChallengeInterface;
 }
